Fix operate-and-get-next skipping command after a multiline command.
When the function operate-and-get-next (C-o) is used on a multi-line
command, if the history is full, the next prompt would show not the
next command but the one after that.

E.g. after the following commands, provided your history is already
full (i.e. already contains $HISTSIZE entries):

  $ echo 'a
  b'
  $ echo c
  $ echo d

going up three places to the two-line echo and hitting C-o would
cause the next prompt to be prefilled with "echo d", not "echo c".

The bug appears to be caused by the extra logic that was added to
set_saved_history to fix this previous bug:
  https://lists.gnu.org/archive/html/bug-bash/2011-09/msg00050.html
Just removing that logic would fix this bug, but naturally reintroduce
that one.

Instead, we can fix both while as a bonus making the code a bit
simpler.  For communicating from operate_and_get_next to
set_saved_history the information of where in the history we want to
be, instead of storing a physical offset like "where_history ()", we
store a logical offset like "where_history () + history_base".  This
allows us to not worry about deducing whether the history got shuffled
or not, because by design the logical offset of a given entry isn't
affected when that happens.

Also add tests for both the previous bug (which occurs when the
multi-line command was read from $HISTFILE) and the present one (which
occurs when it was entered into the current shell session directly).